fix: disable ytdl_hook on macOS/Linux; 0.1.18
- fixed video not starting on macOS/Linux: ytdl_hook ran unconfigured
  against whatever yt-dlp/youtube-dl happened to be on PATH and could
  block the demuxer thread on an unauthenticated probe of the provider
- fixed a couple of clippy lints in the macOS embedding code

+++ src-tauri/src/platform/macos/fullscreen.rs
@@ -59,7 +59,7 @@ pub fn set_simple_fullscreen(app: &AppHandle, on: bool) -> bool {
             // it has to go for the window to cover the screen.
             *WINDOWED_STYLE_MASK
                 .lock()
-                .unwrap_or_else(|e| e.into_inner()) = Some(parent.styleMask().0 as usize);
+                .unwrap_or_else(|e| e.into_inner()) = Some(parent.styleMask().0);
             toggle_style_mask(&parent, NSWindowStyleMask::Titled, false);
 
             let target = screen.frame();
